mes: Try to avoid module name collisions.
The old version of the Guile module system we use puts modules and
variables in the same namespace.  This can cause collisions if you
have two modules: "(a)" and "(a eval)".  When resolving "(a eval)",
the module system looks up the name "eval" in the module "(a)",
which is bound to the core "eval" procedure.  It then assumes it is
a module and crashes.  This change sprinkles some "module?" guards
around to avoid this problem.

* mes/module/mes/guile-module.mes (make-modules-in): Shadow an
existing name if it is not bound to a module.
(resolve-module): Check that 'nested-ref' actually returned a
module.