lib: Fix struct stat's fields for riscv64.
We detected this behavior when building `make` with `tcc` during the
bootstrapping process.  That resulting `make` rebuilt the files, even if
they were just built.

This problem broke the whole Guix bootstrapping process as the
`configure` script was run again in the `build` step because the `make`
program thought it needed a rebuild.  This wouldn't be that much of a
problem if it wasn't that the phase that patches the shebangs on the
generated files is run after `configure` and not after the
"re-configure" that was triggered during the `build` phase.  That made
every generated file that was using `/usr/bin/sh` to explode.

Other architectures (x86_64 is the most suspicious one) may have the
same issue.

NOTE: We used `musl` as a reference, but we have different field names,
which ones should we keep?

    Ours:                              Musl:
      time_t         st_atime;           long st_atime_sec;
      unsigned long  st_atime_usec;      long st_atime_nsec;
      time_t         st_mtime;           long st_mtime_sec;
      unsigned long  st_mtime_usec;      long st_mtime_nsec;
      time_t         st_ctime;           long st_ctime_sec;
      unsigned long  st_ctime_usec;      long st_ctime_nsec;

* include/linux/riscv64/kernel-stat.h (struct stat): Introduce padding
and fix field lengths.