lib: Define va_args for tinycc mob.
In TinyCC (mob) there's a program that generates all the variable
argument control code.  We don't use it in our bootstrapping system so
we need to provide the headers for that ourselves.  We had the support
done for 927, so just removing the version limit makes the support
available for TinyCC (mob) too.

* include/stdarg.h: Remove < 928 version selection for __TINYC__