mes: Remove pmatch-c[ad]r hack.
This commit "fixes" the issue that made us add pmatch-car and
pmatch-cdr, that wasn't a real solution, and made the interpreter way
more complex, but it had no edge cases.  This isn't a real fix either.

The issue appeared in ast-strip-comment because the pmatch pattern used
,cdr and ,car that was effectively rebinding the car and cdr functions,
shadowing the actual cdr and car.

The real fix would be to make the macro hygienic enough so it doesn't
actually shadow anything.  But just changing the binding names makes
mescc work.  A testcase will follow, that covers this lack of hygiene so
we can fix it later better, having a test at hand.

* module/mescc/preprocess.scm (ast-strip-comment): Don't bind car and
  cdr in pmatch.
* mes/module/system/base/pmatch.scm (pmatch-car, pmatch-cdr): Remove and
  use car and cdr instead.
* src/eval-apply.c (cell_vm_eval_pmatch_car, cell_vm_eval_pmatch_cdr):
  Remove.
* src/symbol.c (init_symbols_) [cell_symbol_pmatch_car,
cell_symbol_pmatch_cdr]: Remove.
* include/mes/symbols.h: Likewise.
* src/mes.c (mes_environment): Likewise.
* HACKING: Update.